requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.30
python-dotenv==0.19.2
Werkzeug==2.0.3   # Downgrade to a compatible version
pytest==7.2.0 
requests==2.26.0
prometheus_client==0.16.0
prometheus_flask_exporter==0.23.1